registration/requirements.txt
Django==2.2.7
psycopg2-binary==2.8.6
django-rest-framework==0.1.0
django-rest-registration==0.5.2
djangorestframework==3.9.4
djangorestframework_simplejwt==4.4.0
PyJWT==1.7.1
django-cors-headers==3.7.0
gunicorn==20.1.0
requests==2.26.0
